# docs/openapi/hermes-v1.yaml
openapi: 3.1.0
info:
  title: Hermes Remote Admin API
  version: 1.0.0
  description: |
    Legitimate remote device administration — MVP.
    Admin routes require JWT (2FA verified). Device routes use long-lived device tokens.
servers:
  - url: https://api.example.com/api/v1
    description: Production (replace with your VPS hostname)
...
paths:
...
  /devices/pair:
    post:
      tags: [pairing]
      summary: Pair device with code (no admin JWT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevicePairRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevicePairResponse'
        '400': { description: Invalid or expired code }
  /devices:
    get:
      tags: [devices]
      security: [{ adminBearer: [] }]
      summary: List devices
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/DeviceSummary' }
                  total: { type: integer }
  /devices/me:
    get:
      tags: [devices]
      security: [{ deviceBearer: [] }]
      summary: Current device profile
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceMe'
  /devices/me/heartbeat:
    post:
      tags: [devices]
      security: [{ deviceBearer: [] }]
      summary: Report minimal inventory / liveness
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeartbeatRequest'
      responses:
        '204': { description: Accepted }
  /devices/me/rotate-token:
    post:
      tags: [devices]
      security: [{ deviceBearer: [] }]
      summary: Issue new device token (invalidates previous)
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [device_token]
                properties:
                  device_token:
                    type: string
                    description: Plaintext shown once
...
  /devices/{device_id}/revoke:
    post:
      tags: [devices]
      security: [{ adminBearer: [] }]
      parameters:
        - name: device_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Device revoked }
...
  /files/upload:
    post:
      tags: [files]
      security: [{ deviceBearer: [] }]
      summary: Upload file for a command
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [command_id, file]
              properties:
                command_id: { type: string, format: uuid }
                file:
                  type: string
                  format: binary
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMeta'
  /files/{file_id}/download:
    get:
      tags: [files]
      security: [{ deviceBearer: [] }]
      parameters:
        - name: file_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Binary stream
          content:
            application/octet-stream: {}
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceBearer:
      type: http
      scheme: bearer
      bearerFormat: DeviceToken
  schemas:
...
    DevicePairRequest:
      type: object
      required: [pairing_code, device_name]
      properties:
        pairing_code: { type: string }
        device_name: { type: string, maxLength: 120 }
        public_key: { type: string, nullable: true }
    DevicePairResponse:
      type: object
      required: [device_id, device_token]
      properties:
        device_id: { type: string, format: uuid }
        device_token:
          type: string
          description: Plaintext — store once on device
    DeviceSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        platform: { type: string }
        last_seen: { type: string, format: date-time, nullable: true }
        revoked_at: { type: string, format: date-time, nullable: true }
    DeviceMe:
      allOf:
        - $ref: '#/components/schemas/DeviceSummary'
        - type: object
          properties:
            policy_version: { type: integer }
    HeartbeatRequest:
      type: object
      properties:
        battery_percent: { type: number, nullable: true }
        network_type: { type: string, nullable: true }
        app_version: { type: string, nullable: true }
        os_version: { type: string, nullable: true }
        inventory:
          type: object
          additionalProperties: true
          nullable: true
...
    FileMeta:
      type: object
      required: [id, filename, size_bytes, sha256]
      properties:
        id: { type: string, format: uuid }
        filename: { type: string }
        size_bytes: { type: integer }
        sha256: { type: string }
        command_id: { type: string, format: uuid }
...